Refactor: rename simulated_fe.py
