+++ pyxem/data/__init__.py
@@ -37,7 +37,7 @@ from pyxem.data.simulated_si import (
     si_grains_simple,
     si_rotations_line,
 )
-from pyxem.data.simulation_fe import fe_bcc_phase, fe_fcc_phase, fe_multi_phase_grains
+from pyxem.data.simulated_fe import fe_bcc_phase, fe_fcc_phase, fe_multi_phase_grains
 from pyxem.data._data import (
     au_grating,
     pdnip_glass,
